Move errors to a seperate file
