+++ src/frontend/src/helpers/Services.ts
@@ -12,28 +12,11 @@ import {AbstractCompiler,
         Test,
         CompilerResult,
         CompilerMessage} from "./Compiler/Interface";
-
+import {LoginError} from "./Errors";
 export * from "./Storage/Interface";
 export * from "./Compiler/Interface";
-export {Services, GenericError, LoginError, Connection, DispatchFunction};
+export {Services, Connection, DispatchFunction};
 
-class GenericError extends Error {
-  __proto__: Error;
-    constructor(message?: string) {
-        const trueProto = new.target.prototype;
-        super(message);
-        this.__proto__ = new.target.prototype;
-    }
-}
-
-class LoginError extends GenericError {
-  constructor(message: string,
-              public username?: string,
-              public status?: number,
-              public statusText?: number) {
-    super(message);
-  }
-}
 
 class Connection {
   public wsURI: string;
